# Shortcut for dividing by one